+++ build-prosody.sh
@@ -1,6 +1,7 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 # SPDX-FileCopyrightText: 2024 John Livingston <https://www.john-livingston.fr/>
+# SPDX-FileCopyrightText: 2025 Mehdi Benadel <https://mehdibenadel.com>
 #
 # SPDX-License-Identifier: AGPL-3.0-only