+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Fix #660: don't send headers twice on emoji router errors.
 * Fix shebangs (for NixOS compatibility).
+* Translations updates.
 
 ## 12.0.3